+++ transpiler/main/util/fileutils/fileutils.h
@@ -6,6 +6,7 @@
 char* make_ast_filename(char* filename);
 
 char* make_c_filename(char* filename);
+char* make_h_filename(char* filename);
 
 char* make_gcc_cmd(struct Flags* flags, char* fname_out);